pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "localkin-service-audio"
version = "2.0.0"
description = "Voice AI Platform - Local STT/TTS with Chinese Language Support"
readme = "README.md"
license = "MIT"
authors = [
    {name = "LocalKin Team", email = "info@localkin.dev"},
]
maintainers = [
    {name = "LocalKin Team", email = "info@localkin.dev"},
]
keywords = [
    "speech-to-text",
    "text-to-speech",
    "stt",
    "tts",
    "whisper",
    "voice-cloning",
    "chinese-tts",
    "sensevoice",
    "cosyvoice",
    "local-ai",
    "audio-processing",
    "machine-learning",
    "ai",
    "mcp",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    # Core
    "requests>=2.25.0",
    "click>=8.0.0",           # CLI framework (new in v2.0)
    "rich>=10.0.0",           # Rich terminal output (new in v2.0)
    "pydantic>=2.0.0",
    # Audio
    "sounddevice>=0.4.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "scipy>=1.7.0",
    "numpy>=1.21.0,<1.25.0",
    # STT - Whisper family
    "openai-whisper==20240930",
    "faster-whisper>=1.0.0",
    "pywhispercpp>=1.0.0",
    "ctranslate2>=4.0.0",
    # TTS
    "pyttsx3>=2.90",
    "kokoro>=0.9.2",
    "TTS>=0.22.0",
    # ML
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.21.0",
    "huggingface-hub>=0.17.0",
    "sentencepiece>=0.2.1",
    # Compatibility
    "llvmlite>=0.41.0",
    "numba>=0.58.0",
    # API
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.900",
    "pre-commit>=2.15.0",
    "ruff>=0.1.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.15.0",
]
web = [
    "jinja2>=3.0.0",
    "python-multipart>=0.0.6",
    "websockets>=10.0",
]
tts = [
    "TTS>=0.22.0",
]
gpu = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

# Chinese language models (STT + TTS)
# Note: Install these manually as they may have complex dependencies
# pip install funasr modelscope
chinese = [
    "onnxruntime>=1.15.0",    # For FunASR models
]

# Voice cloning models
# Note: f5-tts and cosyvoice require manual installation
cloning = [
    "onnxruntime>=1.15.0",
]

# MCP server for Claude integration
mcp = [
    "websockets>=10.0",
]

# All new features (v2.0)
all-new = [
    "onnxruntime>=1.15.0",
    "websockets>=10.0",
    "pyyaml>=6.0",
]

[project.scripts]
kin = "localkin_service_audio.cli:main"

[project.urls]
Homepage = "https://localkin.ai/"
Documentation = "https://localkin.ai/"
Repository = "https://github.com/LocalKinAI/localkin-service-audio"
Issues = "https://github.com/LocalKinAI/localkin-service-audio/issues"
Changelog = "https://github.com/LocalKinAI/localkin-service-audio/blob/main/CHANGELOG.md"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["localkin_service_audio*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.setuptools.package-data]
localkin_service_audio = [
    "core/models.json",
    "templates/*",
    "ui/templates/*",
    "ui/static/*",
    "samples/*",
    "README.md"
]

[tool.uv]
dev-dependencies = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.900",
    "pre-commit>=2.15.0",
    "ruff>=0.1.0",
    "transformers<4.50",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["localkin_service_audio"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sounddevice.*",
    "pyttsx3.*",
    "scipy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["localkin_service_audio"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
